requirements.txt
torch
torchvision
transformers
scipy
numpy
Pillow
tqdm
matplotlib
scikit-learn
pyyaml
